requirements.txt
faiss-cpu>=1.7.0
fastapi>=0.100.0
uvicorn>=0.20.0
requests>=2.28.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
numpy>=1.24.0
langchain-openai>=0.1.0
langgraph>=0.2.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
streamlit>=1.30.0
fpdf2>=2.7.0
